+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.9.11
+- Converted to TypeScript
+- Added multilevel logging with colors
+
+# 0.9.10
+- Package updated `electron-prompts-client` in `static/prompt/prompt-client.js`
+
 # 0.9.9
 - Add support and docs for custom prompts
 - Add further support for `input` type Form Elements, including `type="file"`